Separate I2C code in LovyanGFX and twowire driver
MPU6050 and GPIOExtender use twowire i2c driver and LovyanGFX has it's
own implementation. This makes the twowire collide in some way and you
get a 1s delay after each touch event.

To solve this we re-init the twowire with a I2C_0.begin()/end() around
the twowire code in the loop.

This make the speed go down from 713 loops per secund to 470 loops per
second but that should still be ok.